single-touch.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Touch Events tests (based on Matt Brubeck's single-touch tests)</title>
  <meta name="viewport" content="width=device-width">
  <script src="../../../testharness.js"></script>
  <script>
    setup({explicit_done: true});

    // Check a Touch object's atttributes for existence and correct type
    function check_Touch_object (t) {
        test(function() {
           assert_equals(Object.prototype.toString.call(t), "[object Touch]", name + " attribute of type TouchList");
        }, "touch point is a Touch object");
        [
          ["long", "identifier"],
          ["EventTarget", "target"],
          ["long", "screenX"],
          ["long", "screenY"],
          ["long", "clientX"],
          ["long", "clientY"],
          ["long", "pageX"],
          ["long", "pageY"],
        ].forEach(function(attr) {
            var type = attr[0];
            var name = attr[1];

            // existence check
            test(function() {
                assert_true(name in t, name + " attribute in Touch object");
            }, "Touch." + name + " attribute exists");

            // type check
            switch(type) {
            case "long":
                test(function() {
                    assert_equals(typeof t[name], "number", name + " attribute of type long");
                }, "Touch." + name + " attribute is of type number (long)");
                break;
            case "EventTarget":
                // An event target is some type of Element
                test(function() {
                    assert_true(t[name] instanceof Element, "EventTarget must be an Element.");
                }, "Touch." + name + " attribute is of type Element");
                break;
            default:
                break;
            }
       }); 
    }

    // Check a TouchList object's attributes and methods for existence and proper type
    // Also make sure all of the members of the list are Touch objects
    function check_TouchList_object (tl) {
        test(function() {
           assert_equals(Object.prototype.toString.call(tl), "[object TouchList]", name + " attribute of type TouchList");
//           assert_true(tl instanceof TouchList, "tl is a TouchList object");
        }, "touch list is a TouchList object");
        [
          ["unsigned long", "length"],
          ["function", "item"],
          ["function", "identifiedTouch"],
        ].forEach(function(attr) {
            var type = attr[0];
            var name = attr[1];

            // existence check
            test(function() {
                assert_true(name in tl, name + " attribute in TouchList");
            }, "TouchList." + name + " attribute exists");

            // type check
            switch(type) {
            case "unsigned long":
                test(function() {
                    assert_equals(typeof tl[name], "number", name + " attribute of type long");
                }, "TouchList." + name + " attribute is of type number (unsigned long)");
                break;
            case "function":
                test(function() {
                    assert_equals(typeof tl[name], "function", name + " attribute of type long");
                }, "TouchList." + name + " attribute is of type function");
                break;
            default:
                break;
            }
       }); 
       // Each member of tl should be a proper Touch object
       for (var i=0; i < tl.length; i++) {
           check_Touch_object(tl.item(i));
       }
    }

    // Check a TouchEvent event's attributes for existence and proper type
    // Also check that each of the event's TouchList objects are valid
    function check_TouchEvent(ev) {
        test(function() {
           assert_true(ev instanceof TouchEvent, "event is a TouchEvent event");
        }, ev.type + " event is a TouchEvent event");
        [
          ["TouchList", "touches"],
          ["TouchList", "targetTouches"],
          ["TouchList", "changedTouches"],
          ["boolean", "altKey"],
          ["boolean", "metaKey"],
          ["boolean", "ctrlKey"],
          ["boolean", "shiftKey"],
        ].forEach(function(attr) {
            var type = attr[0];
            var name = attr[1];

            // existence check
            test(function() {
                assert_true(name in ev, name + " attribute in " + ev.type + " event");
            }, ev.type + "." + name + " attribute exists");

            // type check
            switch(type) {
            case "boolean":
                test(function() {
                    assert_equals(typeof ev[name], "boolean", name + " attribute of type boolean");
                }, ev.type + "." + name + " attribute is of type boolean");
                break;
            case "TouchList":
                test(function() {
                    assert_equals(Object.prototype.toString.call(ev[name]), "[object TouchList]", name + " attribute of type TouchList");
                }, ev.type + "." + name + " attribute is of type TouchList");
                // Now check the validity of the TouchList
                check_TouchList_object(ev[name]);
                break;
            default:
                break;
            }
       }); 
    }

    function run() {
      var target0 = document.getElementById("target0");
      var target1 = document.getElementById("target1");

      var test_touchstart = async_test("touchstart event received");
      var test_touchmove = async_test("touchmove event received");
      var test_touchend = async_test("touchend event received");
      var test_mousedown = async_test("Interaction with mouse events");

      var touchstart_received = false;
      var touchmove_received = false;
      var touchend_received = false;
      var invalid_touchmove_received = false;
      var touchstart_identifier;

      on_event(target0, "touchstart", function onTouchStart(ev) {
        ev.preventDefault();

        touchstart_received = true;

        // Check ordering
        test_touchstart.step(function() {
          assert_false(touchmove_received, "touchstart precedes touchmove");
          assert_false(touchend_received, "touchstart precedes touchend");
        });
        test_touchstart.done();
        test_mousedown.done(); // If we got here, then the mouse event test is not needed.
        check_TouchEvent(ev);

        test(function() {
          assert_equals(ev.touches.length, 1, "One touch point.");
          assert_equals(ev.changedTouches.length, 1, "One changed touch point.");
          assert_equals(ev.targetTouches.length, 1, "One target touch point.");
        }, "touchstart TouchList lengths are correct.");

        var t = ev.touches[0];
        var ct = ev.changedTouches[0];
        var tt = ev.targetTouches[0];
        
        touchstart_identifier = t.identifier;
        test(function() {
          assert_equals(ct.identifier, touchstart_identifier, "changedTouches identifier matches.");
          assert_equals(tt.identifier, touchstart_identifier, "targetTouches identifier matches.");
        }, "Touch identifiers are consistent.");

        test(function() {
          assert_equals(ev.touches.identifiedTouch(touchstart_identifier), t, "touches.identifiedTouch is correct.");
        }, "identifiedTouch");
      });

      on_event(target0, "touchmove", function onTouchMove(ev) {
        ev.preventDefault();

        if (touchmove_received)
          return;
        touchmove_received = true;

        test_touchmove.step(function() {
          assert_true(touchstart_received, "touchmove follows touchstart");
          assert_false(touchend_received, "touchmove precedes touchend");
        });
        test_touchmove.done();

        check_TouchEvent(ev);

        test(function() {
          assert_equals(ev.touches.length, 1, "One touch point.");
          assert_equals(ev.changedTouches.length, 1, "One changed touch point.");
          assert_equals(ev.targetTouches.length, 1, "One target touch point.");
        }, "touchmove TouchList lengths are correct.");

        test(function() {
          assert_equals(ev.touches[0].identifier, touchstart_identifier, "Touch identifier matches.");
          assert_equals(ev.changedTouches[0].identifier, touchstart_identifier, "Changed touch identifier matches.");
          assert_equals(ev.targetTouches[0].identifier, touchstart_identifier, "Target touch identifier matches.");
        }, "touchmove identifier matches touchstart identifier.");

      });

      on_event(target1, "touchmove", function onTouchMove(ev) {
        invalid_touchmove_received = true;
      });

      on_event(window, "touchend", function onTouchEnd(ev) {
        touchend_received = true;

        test_touchend.step(function() {
          assert_equals(ev.target, target0, "touchend is dispatched to the original target");
          assert_true(touchstart_received, "touchend follows touchstart");
          assert_true(touchmove_received, "touchend follows touchmove");
          assert_false(invalid_touchmove_received, "touchmove dispatched to correct target");
        });
        test_touchend.done();

        check_TouchEvent(ev);

        test(function() {
          assert_equals(ev.touches.length, 0, "Zero touch points.");
          assert_equals(ev.changedTouches.length, 1, "One changed touch point.");
          assert_equals(ev.targetTouches.length, 0, "Zero target touch points.");
        }, "touchend TouchList lengths are correct.");

        var t = ev.changedTouches[0];

        test(function() {
          assert_equals(t.identifier, touchstart_identifier, "changedTouches identifier matches.");
        }, "touchend identifier matches.");

        done();
      });

      on_event(target0, "mousedown", function onMouseDown(ev) {
        test_mousedown.step(function() {
          assert_true(touchstart_received,
            "The touchstart event must be dispatched before any mouse " +
            "events. (If this fails, it might mean that the user agent does " +
            "not implement W3C touch events at all.)"
          );
        });
        test_mousedown.done();

        if (!touchstart_received) {
          // Abort the tests.  If touch events are not supported, then most of
          // the other event handlers will never be called, and the test will
          // time out with misleading results.
          done();
        }
      });
    }
  </script>
  <style>
    div {
      margin: 0em;
      padding: 2em;
    }
    #target0 {
      background: yellow;
      border: 1px solid orange;
    }
    #target1 {
      background: lightblue;
      border: 1px solid blue;
    }
  </style>
</head>
<body onload="run()">
  <h1>Touch Events: single-touch tests</h1>
  <div id="target0">
    Touch this box with one finger (or other pointing device)...
  </div>
  <div id="target1">
    ...then drag to this box and lift your finger.
  </div>
  <div id="log"></div>
</body>
</html>
